msemu: Implement slot_map and dev_map
The dev_map is an array of pointers, one for each device
available in the Mailstation. Rather than separate ptrs
for each device buffer, they are in this array. An enum
is used to easily look up a device based on its number.
These number match the device numbers the Mailstation uses.

The slot_map is also an array of pointers, but is intended
to be a lookup table for the whole address space. Each of
the 16 KiB slots are mapped to one of the ptrs in the array.
These point to the start of the configured page for the set
device. Any time the slot4 or slot8 page or dev changes, the
respective slot_map element is recalculated to point to the
correct page offset in the respective dev_map.

The mread and mwrite functions were overhauled to use this
new setup. This greatly reduced their complexity.