msemu: Convert all NMI calls to INT calls
Also add proper return for intread callback.

It was found that the real hardware for sure implements INT, but
it is now not known when or if NMI is ever used. The external NMI
pin on the Mailstation CPU is tied high, there is no RETN call in
the firmware, but there is also careful code practices around the
0x0038 INT handler and the 0x0066 NMI handler in the stock firmware.