msemu: Fix #33, allow power cycling of MS
Commit 04267bb2e13e9832 broke the ability to turn the power off
and back on again of the emulator. The switch from NMI to INT
caused a subtle bug that actually prevented proper power off.

The real hardware, at power off, does a HALT instruction and cuts
power to a large amount of the peripherals. In this emulation
state, NMI is possible but INT is not. This caused an infinite
loop of HALT instructions once the switch to INT was made. This
was resolved by instead checking the state of the instruction to
ensure the Z80 emulation finished executing an instruction and not
just a prefix, rather than checking if INTs are possible.

This may not be a truly accurate emulation at this point, however,
and could warrant further investigation in the future to better
emulate real hardware