msemu: Interrupt hack for IM 2 modes
There seems to be something that triggers an interrupt outside
of what is currently handled by msemu. This hack enables FyOS v0.1
which uses IM 2 and 256 bytes of a constant value to be combined
with I register to make a jump table that will always jumpt to the
same address for a custom interrupt handler.