ms_state: basic savestate
Unfortunately this is currently dead in the water thanks to z80ex.
The library doesn't let outside world know the contents of the
Z80EX_CONTEXT struct so its not readily possibe for us to save
its state.

Could push a PR to add a snapshot feature. Could integrate z80ex
in to this project directly and add support needed. Who knows.