io: Expand IO buffer to prep for real IO handling
To get ready for real hardware handling, set up IO to have the
simulated IO regs as well as interacting with other real hardware.

Turns the ms->io buffer in to a struct that can be expanded in
the io context without the rest of the code having to worry about
it.

This should also allow us to push some of the other hardware
handling that exists now, like RTC and keyboard, in to the io
context if that is deemed appropriate.